Reorganizing app containers
Reorganizing all components located at containers directory according to their type.
