+++ frontend/src/containers/auth/Register.js
@@ -2,8 +2,8 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 
 import Form from "./Form";
-import Layout from "../components/Layout";
-import { register } from "../store/actions/auth";
+import Layout from "../../components/Layout";
+import { register } from "../../store/actions/auth";
 import RedirectAuthed from "./RedirectAuthed";
 
 class Register extends Component {
